requirements.txt
numpy
scipy
pandas
tqdm
pyyaml
soundfile
librosa
torch